+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-## master
+## 3.0.5
 
 - Fix issue where dropdown is hard to close on Firefox (#194, thanks to @dmarchuk).
 - Filter selected children from results (#201, thanks to @ne0guille).